+++ modules/logging/ui/src/sap/commerce/toolset/logging/actionSystem/CxDeleteCustomLogTemplateAction.kt
@@ -26,7 +26,7 @@ import com.intellij.openapi.ui.Messages
 import com.intellij.util.asSafely
 import sap.commerce.toolset.HybrisIcons
 import sap.commerce.toolset.logging.custom.CxCustomLogTemplateService
-import sap.commerce.toolset.logging.selectedNode
+import sap.commerce.toolset.logging.selectedNodes
 import sap.commerce.toolset.logging.ui.tree.nodes.CxCustomLogTemplateItemNode
 
 class CxDeleteCustomLogTemplateAction : AnAction() {
@@ -38,24 +38,30 @@ class CxDeleteCustomLogTemplateAction : AnAction() {
         if (!e.presentation.isVisible) return
 
         val project = e.project ?: return
-
-        val templateNode = e.selectedNode()
-            ?.asSafely<CxCustomLogTemplateItemNode>()
-            ?: return
+        val selectedNodes = customLogTemplateItemNodes(e) ?: return
+        val message = if (selectedNodes.size == 1) "Delete template \"${selectedNodes.first().name}\"?"
+        else "Delete ${selectedNodes.size} templates?"
 
         if (Messages.showYesNoDialog(
                 project,
-                "Delete template \"${templateNode.name}\"?",
-                "Delete Custom Template",
+                message,
+                "Confirm Deletion",
                 HybrisIcons.Log.Action.DELETE
             ) != Messages.YES
         ) return
 
-        CxCustomLogTemplateService.getInstance(project).deleteCustomTemplate(templateNode.uuid)
+        val templateIds = selectedNodes.map { it.uuid }
+
+        CxCustomLogTemplateService.getInstance(project).deleteTemplates(templateIds)
     }
 
     override fun update(e: AnActionEvent) {
-        e.presentation.text = "Delete Template"
+        val selectedNodes = customLogTemplateItemNodes(e) ?: return
+
+        e.presentation.text = if (selectedNodes.size == 1) "Delete Template" else "Delete Templates"
         e.presentation.icon = HybrisIcons.Log.Template.DELETE_CUSTOM_TEMPLATE
     }
+
+    private fun customLogTemplateItemNodes(e: AnActionEvent): List<CxCustomLogTemplateItemNode>? = e.selectedNodes()
+        ?.mapNotNull { it.asSafely<CxCustomLogTemplateItemNode>() }
 }